Create DepthFirstSearch
This file includes DFS traversals recursive and stack based as well as explicit (createTree() function) and implicit (insert() function) insert functions which are the main building block of the code. Because if tree is made then all the traversal functions and key finding function would run otherwise it will never run.